backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# CrewAI and LLM (compatible versions)
crewai==0.28.8
openai>=1.13.3,<2.0.0

# Data Processing
pandas==2.1.4
python-multipart==0.0.6

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1